rename: login 기능 테스트, register 기능 테스트 코드 패키지 이동 및 이름 변경
